+++ backend/src/services/crewService.js
@@ -10,45 +10,45 @@ module.exports = {
     async create(name, about) {
         let imageURL = name.toLowerCase() + "_avatar.png";
         let crew = await knex("crews").where({name}).first();
-        
-        try {
-            if(crew)
-                return {message: "Equipe já existe!"}
-
-            await knex("crews").insert({
-                id: v4(),
-                name, 
-                about,
-                imageURL
-            });
-
-            return {message: "Equipe criada!"}
-        } catch (error) {
-            throw new Error(error.message);
+    
+        if(crew){
+            throw new Error("Equipe já existe!");
         }
+
+        await knex("crews").insert({
+            id: v4(),
+            name, 
+            about,
+            imageURL
+        });
+
+        return {message: "Equipe criada!"}
     },
 
     async update(id, crew) {
-        try {
-            await knex("crews").where({id}).update(crew);
-            return {message: "Equipe atualizada!"}
-        } catch (error) {
-            throw new Error(error.message);
+        let Crew = await knex("crews").where({id}).first();
+        if(!Crew){
+            throw new Error("Equipe não existe!");
         }
+
+        await knex("crews").where({id}).update(crew);
+        return {message: "Equipe atualizada!"}
+        
     },
 
     async delete(id) {
-        try {
-            let confirmation = await knex("crews").where({id}).delete();
 
-            if(confirmation > 1) {
-                return {message: "Equipes deletadas!"};
-            }
-            
-            return {message: "Equipe deletada!"};
-        } catch (error) {
-            throw new Error(error.message);
+        let crew = await knex("crews").where({id}).first();
+        if(!crew){
+            throw new Error("Equipe não existe!");
         }
+
+        let confirmation = await knex("crews").where({id}).delete();
+        if(confirmation > 1) {
+            return {message: "Equipes deletadas!"};
+        }
+            
+        return {message: "Equipe deletada!"};
     },
 
     getCrewByName(crew_name){
